+++ src/tools/__init__.py
@@ -23,3 +23,4 @@ from src.tools.cookbook import (  # noqa: F401
 )
 from src.tools.search import do_search_chats  # noqa: F401
 from src.tools.notes import do_manage_notes  # noqa: F401
+from src.tools.calendar import do_manage_calendar  # noqa: F401
